0.0.1f
Added a separate thread to handle the computation so the UI thread can show progress without freezing up, refactored some stuff, lots of bugfixes and optimizations.

There is currently a bug in the flood fill algorithm that is adding too many pixels to each convex object. Not sure of the cause yet, will investigate further later.